test: update tests for RAMSTKAllocation record.py and table.py
