+++ tests/models/allocation/conftest.py
@@ -1,0 +1,129 @@
+# Third Party Imports
+import pytest
+from mocks import MockDAO
+
+# RAMSTK Package Imports
+from ramstk.models import RAMSTKAllocationRecord
+
+
+@pytest.fixture()
+def mock_program_dao(monkeypatch):
+    _allocation_1 = RAMSTKAllocationRecord()
+    _allocation_1.revision_id = 1
+    _allocation_1.hardware_id = 1
+    _allocation_1.availability_alloc = 0.0
+    _allocation_1.env_factor = 1
+    _allocation_1.goal_measure_id = 1
+    _allocation_1.hazard_rate_alloc = 0.0
+    _allocation_1.hazard_rate_goal = 0.0
+    _allocation_1.included = 1
+    _allocation_1.int_factor = 1
+    _allocation_1.allocation_method_id = 1
+    _allocation_1.mission_time = 100.0
+    _allocation_1.mtbf_alloc = 0.0
+    _allocation_1.mtbf_goal = 0.0
+    _allocation_1.n_sub_systems = 1
+    _allocation_1.n_sub_elements = 1
+    _allocation_1.parent_id = 0
+    _allocation_1.percent_weight_factor = 0.0
+    _allocation_1.reliability_alloc = 1.0
+    _allocation_1.reliability_goal = 0.999
+    _allocation_1.op_time_factor = 1
+    _allocation_1.soa_factor = 1
+    _allocation_1.weight_factor = 1
+
+    _allocation_2 = RAMSTKAllocationRecord()
+    _allocation_2.revision_id = 1
+    _allocation_2.hardware_id = 2
+    _allocation_2.availability_alloc = 0.0
+    _allocation_2.env_factor = 1
+    _allocation_2.goal_measure_id = 1
+    _allocation_2.hazard_rate_alloc = 0.0
+    _allocation_2.hazard_rate_goal = 0.0
+    _allocation_2.included = 1
+    _allocation_2.int_factor = 1
+    _allocation_2.allocation_method_id = 1
+    _allocation_2.mission_time = 100.0
+    _allocation_2.mtbf_alloc = 0.0
+    _allocation_2.mtbf_goal = 0.0
+    _allocation_2.n_sub_systems = 1
+    _allocation_2.n_sub_elements = 1
+    _allocation_2.parent_id = 1
+    _allocation_2.percent_weight_factor = 0.0
+    _allocation_2.reliability_alloc = 1.0
+    _allocation_2.reliability_goal = 0.9999
+    _allocation_2.op_time_factor = 1
+    _allocation_2.soa_factor = 1
+    _allocation_2.weight_factor = 1
+
+    _allocation_3 = RAMSTKAllocationRecord()
+    _allocation_3.revision_id = 1
+    _allocation_3.hardware_id = 3
+    _allocation_3.availability_alloc = 0.0
+    _allocation_3.env_factor = 1
+    _allocation_3.goal_measure_id = 1
+    _allocation_3.hazard_rate_alloc = 0.0
+    _allocation_3.hazard_rate_goal = 0.0
+    _allocation_3.included = 1
+    _allocation_3.int_factor = 1
+    _allocation_3.allocation_method_id = 1
+    _allocation_3.mission_time = 100.0
+    _allocation_3.mtbf_alloc = 0.0
+    _allocation_3.mtbf_goal = 0.0
+    _allocation_3.n_sub_systems = 1
+    _allocation_3.n_sub_elements = 1
+    _allocation_3.parent_id = 1
+    _allocation_3.percent_weight_factor = 0.0
+    _allocation_3.reliability_alloc = 1.0
+    _allocation_3.reliability_goal = 0.99999
+    _allocation_3.op_time_factor = 1
+    _allocation_3.soa_factor = 1
+    _allocation_3.weight_factor = 1
+
+    DAO = MockDAO()
+    DAO.table = [
+        _allocation_1,
+        _allocation_2,
+        _allocation_3,
+    ]
+
+    yield DAO
+
+
+@pytest.fixture(scope="function")
+def test_attributes():
+    yield {
+        "revision_id": 1,
+        "hardware_id": 1,
+        "availability_alloc": 0.0,
+        "env_factor": 1,
+        "goal_measure_id": 1,
+        "hazard_rate_alloc": 0.0,
+        "hazard_rate_goal": 0.0,
+        "included": 1,
+        "int_factor": 1,
+        "allocation_method_id": 1,
+        "mission_time": 100.0,
+        "mtbf_alloc": 0.0,
+        "mtbf_goal": 0.0,
+        "n_sub_systems": 1,
+        "n_sub_elements": 1,
+        "parent_id": 7,
+        "percent_weight_factor": 0.0,
+        "reliability_alloc": 1.0,
+        "reliability_goal": 0.999,
+        "op_time_factor": 1,
+        "soa_factor": 1,
+        "weight_factor": 1,
+    }
+
+
+@pytest.fixture(scope="function")
+def test_recordmodel(mock_program_dao):
+    """Get a record model instance for each test function."""
+    dut = mock_program_dao.do_select_all(RAMSTKAllocationRecord, _all=False)
+
+    yield dut
+
+    # Delete the device under test.
+    del dut
